vibe-synth/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
redis==5.0.1
google-generativeai==0.3.1
textblob==0.17.1
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
asyncio-mqtt==0.16.1